bloom: healObject to mark a path dirty only for dangling objects
Path is marked dirty automatically when healObject() is called, which is
wrong. HealObject() is called during self healing and this will lead to
the increase of the false positive result of the bloom filter.

Also move NSUpdated() from renameData() and call it directly in
CompleteMultipart and PutObject, this is not a functional change but
it will make it less prone to errors in the future.