Implement inspect data API v2
The old inspect format is still supported

The new format supports data encryption with a public key, that will be
received from madmin-go/mc.

The real data encryption key is still random, and it will be encrypted with
the provided RSA public key. The encrypted key will be shipped in the
archive as key.enc