Quick initialized of bucket target notifications
The initialization of the bucket target notification is lazy now. It is
initialized with FlushQueueStore() called in bucket target loading or
in the first Send().

Loading the configuration of bucket notifications only fails when the
queue store in not properly configured. It doesn't ping the target
notification and checks if it is online or not.

If one event needed to be sent when the target notifications is not
fully initialized yet, the event is saved in a channel. The channel will
be freed when all targets are initialized.

The channel is needed because currently the initialization of the object
layer happens before the initialization of the configuration.