Save resync status in the backend using a last update timestamp
Currently there is a short time window where the code is allowed to save
the status of a replication resync. Currently, the window is
`now.Sub(st.EndTime) <= resyncTimeInterval`. Also any failure to write
in the backend disks is not retried.

Refactor the code a little bit to rely on the last timestamp of a
successful write of the resync status of any given bucket in the backend
disks.