persist the non-default creds in config
non-default creds should persist to disk
and we must detect that they are not set
to "default" credentials by mistake.

Print a warning indicating that the
deployment is now using default credentials.

NOTE: we cannot stop the service instead
we can simply warn the users at this point
in time.