Fix ignoring odirect in case of erasure single disk
The current code does not properly ignore o_direct error when
initializing format.json, it fails to detect Write() error as invalid
argument because odirect library Write() is returning *errors.errorString
error and not syscall.Errno

The code also assumes the o_direct is supported if format.json already
exists, which is wrong.

The only way to fix this is to check O_DIRECT support in each server
start.