validate correct ETag for the parts sent during CompleteMultipart
Current code blindly assumes if part number matches then its
corresponding etag is also same for encrypted parts, this
is a problem of incorrect input validation leading to a situation
where a client can construct a final object with range of those
are not the same as intended for.

This is rare but it is a plausibility, also going through all
parts 10,000 parts 10,000 times over is an expensive operation
to be done at the handler layer instead do this at the lowest
layer.