Add Data-Dir as criteria to pick the quorum FileInfo
With the replication feature, modtime alone is not reliable enough to
pick a FileInfo{} which has a quorum in the disks. The reason is that
replication sets ModTime when replicating objects. Add Data-Dir criteria
for versions with associated data (not delete marker) to pick FileInfo{}
with correct related info, such as EcN, EcM and other informations.