Tweak objectQuorumFromMeta() to search for parity quorum
In a replication setup, we can have an object REPLICA with the same
modtime in different disks but with different internal information such
as erasure coding information. Currently objectQuorumFromMeta() can pick
a wrong FileInfo{} because modtime was only used as a criteria to pick
the valid FileInfo but that is not always true anymore.

This PR will just pick the parity number that has quorum among disks.