Clean entire tmp-old on restart
In cases where the server is restarted before the previous run had finished would leak the data.

Do a removeAll on the entire folder. This may clash a bit, but will ensure that the path is cleaned on every startup. There is no logging in the subpaths, so it shouldn't be noisy.

To mitigate this somewhat we only do cleanup when health checks are enabled and for the non-rest instances.