allow versions to spread across pools without pool locality
This PR adds a new change where versions for "an" object
to be spread across pools, this allows for versions to be
now spread across pools to provide

- better space management when each version can be bigger
- create newer versions on newer pools to avail better
  responsiveness
- and allows for lesser data movement when older pools
  need to be decommissioned

This PR additionally also optimizes reading xl.meta for
multipart uploads, to avoid multiple lock requests to
read the same content. This should essentially increase
overall performance for multipart requests in multi-pooled
setup.