Distribute replication ops based on names
Background:

Replication workers can end up fighting over the same object locks when many versions of an object are processed at the same time.

This leads to huge resource use and general slowdown of the replication.

Solution:

Instead of having a single channel shared across many workers we give each worker a channel and distribute objects based on a hash of their bucket+object key.

Objects are re-distributed when the worker count is re-scaled, but that should be a rather rare occurrence.