Fix replication deadlocks
When auto-scaling the worker the operation would deadlock.

An Rlock would be held while `ResizeWorkers` would try to grab a write lock. This will deadlock the entire ReplicationPool.

Also make sure that atomic int32 values are aligned to avoid potential 32 bit crashes.

Fixed in #16083 as well, but this is cleaner to backport.