fix: skip free version ID and marker in metadata equality
It so happens that we previously had a bug where an object
might be created with free-version-id (incorrectly)

where it is not supposed to since we explicitly reject
such an occurrence, there was an issue in the past
that would end up preserving tier-versionid for a
regular object.

skip checks for metadata equality by skipping these keys
as they are not supposed to be present via a PUT object.