kms: fix possible deadlock due to nested RLock calls.
This commit fixes a deadlock issue caused by concurrent
`sync.RWMutex.Lock` and nested `sync.RWMutex.RLock` calls.

A `sync.RWMutex` will block subsequent `RLock` callers once
a `Lock` blocks. It does this to ensure that the `Lock` caller
evetually gets the lock and does not starve due to `RLock` calls.

```
RLock locks rw for reading.

It should not be used for recursive read locking;
a blocked Lock call excludes new readers from acquiring the lock.
See the documentation on the RWMutex type.
```
Ref: https://pkg.go.dev/sync#RWMutex.RLock

The previous code violated this by first requiring a `RLock` at the
beginning of `DecryptAll` and then calling `Decrypt` - which itself
requests an `RLock`.

This can cause a deadlock when a `DecryptAll` calls holds the
`RLock` and a concurrent KES certificate reload tries to update
the KES client with a new certificate, and therefore, requests
a `Lock`.

Now, the `Lock` is blocked - since DecryptAll holds an `RLock`.
Further, the `Lock` call blocks any subsequent `RLock` calls.
However, DecryptAll wants to acquire and release the `RLock`
again to complete.

Now, the `Lock` goroutine waits on the `RLock` goroutine to
complete but also blocks subsequent `RLock` calls and the
`RLock` gorotuine wants to acquire another `RLock`, and therefore,
is blocked on the `Lock` goroutine.

=> Deadlock

This commit fixes this by not nesting `RLock` calls.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>