kms: add support for KES enclaves
This commit adds support for KES enclaves.
An enclave is an isolated space within the
KES server that allows tenant separation.

This commit introduces an optional new env. variable:
```
MINIO_KMS_KES_ENCLAVE=<name>
```
If not set, the KES client will not send any enclave
information. On a stateless KES server, this will
result in the default enclave.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>