s3zip: Limit over-read
Since we don't know the exact end offset of a file due to headers, we specified a read until the end-of-the-stream.

This would cause the remainder of the current parts to be read by "streamingBitrotReader" in Close()

For larger parts that is a very significant overhead. We allow 64KB for a single file header and only overread by that amount.

Warp benchmark, 1 object 10000 files inside, each 1MiB, 1GiB part size:

```
λ warp cmp warp-zip-before.csv.zst warp-zip-after.csv.zst
507 operations loaded... Done!
22572 operations loaded... Done!
-------------------
Operation: GET
Operations: 507 -> 22572
Duration: 46s -> 1m0s
* Average: +4557.94% (+368.1 MiB/s) throughput, +4557.94% (+368.1) obj/s
* Requests: Avg: -2.3841s (-98%), P50: -2.2584994s (-98%), P99: -7.3977074s (-99%), Best: -20.3383ms (-41%), Worst: -8.6780461s (-99%)
* TTFB: Avg: -2.06595s (-98%), P50: -1.9913357s (-98%), P99: -5.1238974s (-99%), Best: -21.4768ms (-45%), Worst: -8.6802819s (-99%)
* Fastest: +2894.43% (+371.3 MiB/s) throughput, +2894.42% (+371.3) obj/s
* 50% Median: +4579.98% (+369.3 MiB/s) throughput, +4579.97% (+369.3) obj/s
* Slowest: +5997.88% (+357.4 MiB/s) throughput, +5997.87% (+357.4) obj/s
```