vectorize cluster wide calls such as bucket operations
This PR is a re-implementation of the following calls

 - MakeBucket()
 - DeleteBucket()
 - GetBucketInfo()

into peer-level calls, this is needed to ensure that
we start reducing the number of connections needed to
perform bucket operations - that are always cluster-wide,
with a high number of disk setups, it is important that
we reduce the state of open connections between hosts.

New connections incur costs, that can in turn put a lot
of syscall() load on the system, this PR should reduce
the number of connections by order of magnitude.