Do not return an error when version disparity is detected
This is only for S3 mulipart upload. Currently when a there is
high level of object's versions disparity, the S3 API returns 503.

The reason is that renameData() returns errFileNeedsHealing and it is
not ignored in CompleteMultipartUpload.

Fix the issue and modify renameData() to return a boolean instead of an
error to indicate that an object needs be healed.