Do not save credentials in config.json
Currently, credentials are stored in .minio.sys/config/config.json for
historical reasons, but it is not documented it is the case, neither it
is possible to modify the access/secret keys in config.json

Also it is documented that users need to define MINIO_ROOT_USER and
MINIO_ROOT_PASSWORD in the MinIO environment.

This change can break users who did never define MINIO_ROOT_USER and
MINIO_ROOT_PASSWORD. In that case, the credentials will be
minioadmin:minioadmin, which means this is not a serious deployment.

This can also break users that set MINIO_ROOT_USER and
MINIO_ROOT_PASSWORD with a custom access and secret keys temporarly and
then removed those env variables. But again, this is not a supported use
case since a long time.

Therefore removing credentials subsystem from MinIO configuration.