Heal once when multiple disks in one erasure set are recreated
An example is one node which contains two or more disks belonging to the
same erasure set is recreated. In that case, healing will heal disks
belonging to the same erasure set one by one.

This PR defines a new uuid called HealID which is generated during
HealFormat and saved in the newly formatted disks in .healing.bin

Once the server detects a new disk that needs to be healed and once it
finishes healing, it will search for all disks that have .healing.bin
with the same HealID and remove .healing.bin from them.