s3: Return XMinioInvalidObjectName if the object contains null char
GetObject() on a object name with a null character will return the drive
does not support O_DIRECT though it does. The reason is that the code
does not know for sure why a syscall returns EINVALID and it converts it
to the unsupported O_DIRECT error, which is confusing.

To avoid this situation, just consider \0 as an invalid object character
though S3 supports it, but \0 will never work with filesystems anyway.