Hide objects/delete-markers that are marked to be removed
Add a special flag to versions that are marked for deletion but not
replicated it yet. This flag will be used to hide from listing.

This will only applicable for for new delete replication. Current versions
that fail to be replicated will still be visible, however this should not be
an issue a non replicate delete will succeed in the future after fixing
the underlying issue.