Properly restart fresh disk healing when failed in some places
globalBackgroundHealState.getHealLocalDiskEndpoints() returns disks that
need to be healed, but not the disks that are currently in a healing
process.

Before starting to heal a disk, this latter is marked in the memory as
currently healing. However when there is a healing issue, the code
forgot to mark the disk as non healing anymore, so it never picked again
in the next loop.

The workaround is to restart the node since the state in the memory is
wrong. However the commit will make it right by removing the healing
state of the disk when healing fails for any reason.