Use async healing in PutObject call
PutObject can heal up to two versions when it finds a disparency
between the list of versions, however the heal object call is
sync and holds a lock as well. Call healing in a goroutine instead.

Also, do not account disks returning an error (offline for example) when
trying to detect if there is a versions disparency for an object during
PutObject.