Fix incorrect anonymization of drive endpoint
When the endpoints are not provided in ellipses style in the server
command line, anonymization of drive endpoints was not happening
correctly and was sometimes was adding server indices higher than the
number of servers, creating a misleading impression that the health
diagnostics is reporting a higher number of servers.

Fixed by ensuring correct server index in the anonymized endpoint.