Better error when setting up replication with a service account alias
Creating a replicator service account requires having a parent user. In
this case, the parent user will be the mc alias access-key or the
console access-key and it cannot be a service account since the code
does not support creating a sub service account for an existing service
account.

This commit will just show a better error message in that case.