Fix truncated response on deleted replicated objects
When an object version is deleted it is marked as a "FreeTierVersion" pending replication of the delete.

These will be excluded from listing results since they are deleted, but pending replication.

The issue is that when 1000 keys are requested, we request 1000 keys from each disk, but don't expect all of them to be filtered out. Since all are filtered out, the listing returns as if it has completed.

Do not count all free versions as a returned object when listing with limits.