feat: allow offline disks on a fresh start
We never used to allow offline disks during fresh
setup, but that is not necessary anymore. The reason
being we expect the order of disks to be exactly same
once started.

This new change simply allows starting MinIO with
some offline disks, it still refuses to start if we
do not have enough disks online, i.e

quorum=(len(disks)/2+1)