support if-match/if-none-match with s3 uploads
This PR supports optimistic concurrency via
if-match, if-none-match HTTP headers with s3
uploads.

A client that is performing an upload can send
the original ETag together with a conditional header
to ensure that an update will only occur if a certain
condition has been met. For example, if the If-Match
header is specified, MinIO verifies that the value
of the ETag specified in the update request is the
same as the ETag for the object being uploaded.