fix: scanner in healing cycle must use actual size
HealObject() was returning the size of the object
on disk, instead of the actual size of the object.

On compression enabled setups the final size
resulted in lower count once the healing cycle
was complete causing usage to go lower and
higher at various intervals. This PR fixes the
behavior properly.