heal: Do not mark buckets as done when there is no online disks
Some old code marks a bucket as healed when it is unable to get
any non healing disks in the erasure set.

It is not clear what that code point. Remove it - the healing will
continue healing the remaining data but error out at the end, to trigger
another healing round.