Fix healing deadlocks and ordering
We share respCh between all concurrent calls of a healSequence.
It it sent to a random gorourine waiting that gets unblocked by healResult.

Adding nowait in #16433 made this deadlock, since it was "eating" healResults from other goroutines and they would block until they randomly get a result.
As more goroutines are leaking this gets increasingly rare.

Changes: Each task creates its own result channel. We make it size 1, so a result can always be written. We stop waiting for a result when `noWait` is set.