Fix broken pipe msg from go tool in CI for FIPS
`grep -q` quits after first match, and go tool reports "broken pipe"
because it can no longer write its output to stdout - it looks like the
test is not for the FIPS build binary is not working but it actually is.

To stop the error message, we pipe into grep twice.