Add lazy init of audit logger
If the http endpoint is down when starting, we check if it comes back every second.

We queue audit messages until the queue is full when offline. These will be sent if we should get back online.

Discovered a bazillion races in this, when calling Init and Cancel.

We should now correctly flush the queue and not risk "send on closed channel" panics.

Replaces #16835