Set tcp user timeout to clean sockets with data in the buffer
TCP Keep Alive is not enough to kill connections with dead clients. The
reason is that TCP keep alive does not work when there is data in the
socket buffer waiting to be transmitted. To fix this, TCP_USER_TIMEOUT
needs to be set. It kills connections with non received ACKs in many
connection stages.