trace: Fix <unknown> func name of requests rejected by max clients
Currently, when a request is rejected because it was queued for a long
 time (requests_deadline), the trace event does not appear when filtered
against S3 calls. However we can see the request when running 'mc admin
trace -a', the function name will be '<unknown>'

The function name is set by 'unknown' because it did not reach the S3
handler.

Fix this by setting the trace func name in maxClients(), since
maxClients() is only called for S3 requests.

Also, when the request is canceled, return 499 status code instead of
200 (the default one)