fix: inlined objects don't need to honor long locks
getObjectNInfo() current holds the lock until all
the inlined data is written to disk.

It is sufficient to verify xl.meta and its quorum
validity under quorum, once that is achieved we
do not ever have to hold the lock further.

Currently clients speed would dictate the time
the lock is held for READ operation on MinIO,
this is not needed and we can make this lock-free
since the inlined data is already obtained in
a single call.