Support persistent queue store for loggers
This commit adds the queue store persistence support for MinIO loggers. This
enables persisting the logs to the configured queue store which will be
replayed by the store replay mechanism. This is similar to queue store
support for bucket notification targets.

This will make sure the logs will not be missed when the logger targets
are down. The logs will be persisted in such cases and will be replayed
once the target is up.