allow root user to be disabled via config settings
root user is super admin in almost all the deployments
however once the cluster is setup, this account is
not necessarily needed for any future operations of
the cluster.

Once users have created admins with policies, it is
sufficient for us to disable root users henceforth.

any other admin can re-enable root if needed, and
environment values can override in situations where
root needs to be re-enabled back.