Use small blocks up to 64KB
From https://github.com/minio/minio/discussions/17103 - files up to 1MB would be read in 32KB blocks. Set that threshold to 64KB, so at most there are 2 reads.

Remove the unneeded wrapper in readAllData when we are not using O_DIRECT - we are reading directly into a single buffer, so there is no need for an intermediate buffer. This will save a memcopy in most cases.

Either way, we now quickly reject on unsupported platforms and avoid the copying there.