Check for only network errors in audit webhook for reachability
Currently, If the webhook responds with 500 response code, we consider
the audit target to be down and ping repeateadly to initialize the target. Instead,
we need to just consider actual target reachability error. Just receiving a response
from the webhook is a good enough check for reachability.

This commit also addresses the review comments from https://github.com/minio/minio/pull/17121