Remove older policy attach behavior for LDAP
This removes previously deprecated behavior.

`mc admin policy attach` will now raise an error when the user does not
exist. Previously it was allowed to have a non-existent user to attach a
policy to support setting policies for LDAP authenticated users.

Now, we have `mc idp ldap policy attach` for this purpose which
validates if the user exists in the configured LDAP directory.