Add an option to make bucket notifications synchronous
With the current asynchronous behaviour in sending notification events
to the targets, we can't provide guaranteed delivery as the systems
might go for restarts.

For such event-driven use-cases, we can provide an option to enable
synchronous events where the APIs wait until the event is successfully
sent or persisted.

This commit adds 'MINIO_ENABLE_SYNC_EVENTS' env which when set to 'true'
will enable sending/persisting events to targets synchronously.