Revert "heal: Update object parity with the latest configured SC (#17187)"
This reverts commit e2b7a08c1024659c8119c4b8819f617bfbb92de2.

Healing while updating the parity can corrupt the object in a unstable system.

Before, healing reverts the old state of an outdated disk and not changing
anything in the other disks that do not need healing, so it is acceptable that healing
fails to write in some outdated disks due to any reason, since this will not make
things worse anyway.

However, healing with the parity update enabled will change the state of the object in all
disks and this can be an issue if the new state is propagated in some disks but not in all
of them.

Hence, reverting the commit. Also it doesn't seem there is any solid solution in an unstable cluster
without making much larger involved changes.