Optimize DeleteObject API to avoid extra GetObjectInfo call
on the replicating side. For receiving side, it is just a regular
DeleteObject call.

Also fixing a corner case where version being purged is absent on target (either due to replication not yet complete or target version already deleted in a one-way replication or while replication was disabled). In such cases, mark version purge complete.