interpret empty version as `null` to evolve with versioning configuration
empty versions such as `null` are empty and `null` depending
on the bucket configuration. This can be problematic for
count(versions) calculation as we may skip valid versions.

We may also skip passing correct versionId in ILM via
scanner. Always treat buckets as versioned to ensure that
un-versioned bucket versions are treated as `null`.