don't error when asked for 0-based range on empty objects
In a reverse proxying setup, a proxy in front of minio may attempt to
request objects in slices for enhanced cache efficiency. Since such a
proxy cannot have prior knowledge of how large a requested resource is,
it usually sends a header of the form:

        Range: 0-$slice_size

... and, depending on the size of the resource, expects either:

- an empty response, if $resource_size == 0
- a full response, if $resource_size <= $slice_size
- a partial response, if $resource_size > $slice_size

Prior to this change, minio would respond 416 Range Not Satisfiable if a
client tried to request a range on an empty resource. This behavior is
technically consistent with RFC9110[1] – however, it renders sliced
reverse proxying, such as implemented in nginx, broken in the case of
empty files. Nginx itself seems to break this convention to enable
"useful" responses in these cases, and minio should probably do that
too.

[1]: https://www.rfc-editor.org/rfc/rfc9110#byte.ranges