Do not use a chain for S3 tiering to return better error messages
When using a chain provider and all providers do not return a valid
access and secret key, an anonymous request is sent, which makes it hard
for users to figure out what is going on

In case of S3 tiering, when AWS IAM temporary account generation returns
an error, anynomous login will be used because of the chain provider.
Avoid this and use AWS IAM provider directly to get a good error
message.