Fix various poolmeta races
Code huddle with @r-scheele reviewing the code.

There is a fundamental race condition in `newErasureServerPools`, where setObjectLayer is called before the poolMeta has been loaded/populated.

We add a placeholder value to this field, but disable all saving of the value, so we don't risk overwriting the value on disk. Once the value has been loaded or created, it is replaced with the proper value, which will also save.

Also fixes various accesses of `poolMeta` that was done without locks.

We make the `poolMeta.IsSuspended` return false, even if we shouldn't risk out-of-bounds reads any more.